refactor: change test
